Initial Upload
Contains roughly 5-10 hours of work. Includes a system for spawning starts/planets and a rough estimation of the planetary physics in a 2-D plane. Initiliazed with mass/distance/speed of earth and the sun. Was able to achieve orbit with these numbers (with some error that was accounted for). Also added a pause button.